I Crashed :(
By the looks of it, it saved the final part of the zone i was working on thank god. Also saved the template texture i will use for the particle system.

Looks like the work on the actual partticle system and material are what didnt get saved.

With that im gonna go get food cause im starving.