MORE MORE MORE
Made the icons unlit so they dont shine or reflect etc. developed more of the trap hallway adding in a wooden trench area at the end. Added in audio of cracking wood around the level. Started on and did a big chunck of level progress on the exterior of the temple. Added in some sourced jungle images for the outside.